Add `delayMicroseconds`
This isn't a good version of it, but it has the capability of delaying a
lot shorter times than `delay` can. We could use the register too, but
right now this is fine. Really just need to find a better way to get the
necessary assembly and a nice way to measure 16MHz pulses.